backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.43.0
openai==1.59.3
httpx==0.28.1
python-dotenv==1.0.1
apscheduler==3.10.4
thefuzz[speedup]==0.22.1
pydantic==2.10.4
